requirements.txt
google-genai
sarvamai

fastapi
uvicorn
gunicorn
azure-data-tables
pandas

python-dotenv
aiohttp
aiofiles
soundfile
pydantic
starlette
aiomysql